fix(datacell): widen SparseVectorDataCell offset from uint32 to uint64
The current_offset_ field and per-vector offset stored in offset_io_
were uint32_t, limiting total sparse vector data to 4GB before silent
wraparound and data corruption. Widen to uint64_t so the offset can
address up to 16 EB.

Changes:
- current_offset_ member: uint32_t -> uint64_t
- InsertVector local old_offset: uint32_t -> uint64_t
- GetCodesById local offset: uint32_t -> uint64_t
- Resize offset_io_ slot width: sizeof(uint32_t) -> sizeof(uint64_t)

+++ src/datacell/sparse_vector_datacell.h
@@ -80,7 +80,7 @@ public:
         }
         uint64_t io_size = (new_capacity - total_count_) * max_code_size_ + current_offset_;
         this->io_->Resize(io_size);
-        this->offset_io_->Resize(static_cast<uint64_t>(new_capacity) * sizeof(uint32_t));
+        this->offset_io_->Resize(static_cast<uint64_t>(new_capacity) * sizeof(uint64_t));
         this->max_capacity_ = new_capacity;
     }
 
@@ -159,7 +159,7 @@ private:
 
     Allocator* const allocator_{nullptr};
     std::shared_ptr<MemoryBlockIO> offset_io_{nullptr};
-    uint32_t current_offset_{0};
+    uint64_t current_offset_{0};
     uint64_t max_code_size_{0};
     std::mutex current_offset_mutex_;
 };
